Refactor spec/api.ts into domain-specific modules
Split the large api.ts file (~4000 lines) into separate files organized
by domain:

- api/types.ts - Shared Fetch type
- api/status.ts - Status endpoint test
- api/resources.ts - Resource management tests
- api/brains.ts - Brain execution tests
- api/schedules.ts - Schedule management tests
- api/secrets.ts - Secrets management tests
- api/webhooks.ts - Webhook tests
- api/signals.ts - Signal handling tests
- api/pages.ts - Page management tests
- api/bundle.ts - Bundle endpoint tests
- api/index.ts - Barrel file re-exporting all modules

The public API remains unchanged - consumers still import from
@positronic/spec and get the same exports.
