+++ packages/spec/src/api/index.ts
@@ -1,0 +1,9 @@
+export { testStatus } from './status.js';
+export { resources } from './resources.js';
+export { brains } from './brains.js';
+export { schedules } from './schedules.js';
+export { secrets } from './secrets.js';
+export { webhooks } from './webhooks.js';
+export { signals } from './signals.js';
+export { pages } from './pages.js';
+export { bundle } from './bundle.js';
