+++ packages/spec/src/api/types.ts
@@ -1,0 +1,1 @@
+export type Fetch = (request: Request) => Promise<Response>;
