+++ packages/spec/src/index.ts
@@ -133,4 +133,4 @@ export interface PositronicDevServer {
   bulkSecrets(filePath: string): Promise<void>;
 }
 
-export { testStatus, resources, brains, schedules, secrets, webhooks, signals, pages, bundle } from './api.js';
+export { testStatus, resources, brains, schedules, secrets, webhooks, signals, pages, bundle } from './api/index.js';
